WIP Splitting implementation details from public API
