+++ immer/extra/persist/detail/traits.hpp
@@ -1,6 +1,6 @@
 #pragma once
 
-namespace immer::persist {
+namespace immer::persist::detail {
 
 /**
  * Define these traits to connect a type (vector_one<T>) to its pool
@@ -10,4 +10,4 @@ template <class T>
 struct container_traits
 {};
 
-} // namespace immer::persist
+} // namespace immer::persist::detail
